+++ src/main/java/fr/maxlego08/essentials/commands/commands/essentials/CommandEssentialsConvert.java
@@ -3,8 +3,12 @@ package fr.maxlego08.essentials.commands.commands.essentials;
 import fr.maxlego08.essentials.api.EssentialsPlugin;
 import fr.maxlego08.essentials.api.commands.CommandResultType;
 import fr.maxlego08.essentials.api.commands.Permission;
+import fr.maxlego08.essentials.api.convert.Convert;
 import fr.maxlego08.essentials.convert.cmi.CMIConvert;
+import fr.maxlego08.essentials.convert.coinsengine.CoinsEngineConvert;
 import fr.maxlego08.essentials.convert.essentialsx.EssentialsXConvert;
+import fr.maxlego08.essentials.convert.playervaultx.PlayerVaultXConvert;
+import fr.maxlego08.essentials.convert.sunlight.SunlightConvert;
 import fr.maxlego08.essentials.zutils.utils.commands.VCommand;
 
 import java.util.Arrays;
@@ -15,29 +19,46 @@ public class CommandEssentialsConvert extends VCommand {
         super(plugin);
         this.addSubCommand("convert", "conv", "c");
         this.setPermission(Permission.ESSENTIALS_CONVERT);
-        this.addOptionalArg("plugin", (a, b) -> Arrays.asList("CMI", "EssentialsX", "Sunlight"));
+        this.addOptionalArg("plugin", (a, b) -> Arrays.stream(Plugins.values()).map(Enum::name).map(String::toLowerCase).toList());
     }
 
     @Override
     protected CommandResultType perform(EssentialsPlugin plugin) {
 
         String pluginName = this.argAsString(0);
-        if (pluginName.equalsIgnoreCase("CMI")) {
-
-            CMIConvert cmiConvert = new CMIConvert(plugin);
-            cmiConvert.convert(sender);
-
-            return CommandResultType.SUCCESS;
+        Plugins plugins = Plugins.valueOf(pluginName.toUpperCase());
+        var converClass = plugins.getConvertClass();
+        try {
+            var constructor = converClass.getConstructor(EssentialsPlugin.class);
+            Convert convert = constructor.newInstance(plugin);
+            convert.convert(sender);
+        } catch (Exception exception) {
+            exception.printStackTrace();
+            return CommandResultType.SYNTAX_ERROR;
         }
 
-        if (pluginName.equalsIgnoreCase("EssentialsX")) {
-
-            EssentialsXConvert essentialsXConvert = new EssentialsXConvert(plugin);
-            essentialsXConvert.convert(sender);
-
-            return CommandResultType.SUCCESS;
-        }
-
-        return CommandResultType.SYNTAX_ERROR;
+        return CommandResultType.SUCCESS;
     }
+
+    public enum Plugins {
+
+        ESSENTIALSX(EssentialsXConvert.class),
+        CMI(CMIConvert.class),
+        SUNLIGHT(SunlightConvert.class),
+        PLAYERVAULTX(PlayerVaultXConvert.class),
+        COINSENGINE(CoinsEngineConvert.class),
+
+        ;
+
+        private final Class<? extends Convert> convertClass;
+
+        Plugins(Class<? extends Convert> convertClass) {
+            this.convertClass = convertClass;
+        }
+
+        public Class<? extends Convert> getConvertClass() {
+            return convertClass;
+        }
+    }
+
 }