+++ src/main/java/fr/maxlego08/essentials/convert/cmi/CMIConvert.java
@@ -1,6 +1,7 @@
 package fr.maxlego08.essentials.convert.cmi;
 
 import fr.maxlego08.essentials.api.EssentialsPlugin;
+import fr.maxlego08.essentials.api.convert.Convert;
 import fr.maxlego08.essentials.api.home.Home;
 import fr.maxlego08.essentials.storage.database.repositeries.UserEconomyRepository;
 import fr.maxlego08.essentials.storage.database.repositeries.UserHomeRepository;
@@ -20,7 +21,7 @@ import java.io.File;
 import java.util.ArrayList;
 import java.util.List;
 
-public class CMIConvert extends ZUtils {
+public class CMIConvert extends ZUtils implements Convert {
 
     private final EssentialsPlugin plugin;
 
@@ -28,6 +29,7 @@ public class CMIConvert extends ZUtils {
         this.plugin = plugin;
     }
 
+    @Override
     public void convert(CommandSender sender) {
 
         message(sender, "&fStart convert &7CMI");
@@ -103,7 +105,7 @@ public class CMIConvert extends ZUtils {
             try {
                 String[] parts = entry.split("[$]|%%|:");
                 String name = parts[0];
-                int index =  parts.length == 10 ? 4 : parts.length == 9 ? 3 : 2;
+                int index = parts.length == 10 ? 4 : parts.length == 9 ? 3 : 2;
                 Material material = null;
                 if (parts.length == 9) {
                     try {