+++ changelog.md
@@ -1,5 +1,30 @@
 # Unreleased
 
+# 1.0.0.4
+
+- Added the command `/ess convert PlauerVaultX`, allows converting the database from PlayerVaultX to zEssentials.
+- Added the command `/ess convert Sunlight`, allows converting the database from Sunlight to zEssentials.
+- Added the command `/ess convert CoinsEngine`, allows converting the database from CoinsEngine to zEssentials.
+- Items in player vault will be displayed as a single item.
+- Fixed night vision messages
+- Fixed duplicate lines [#33](https://github.com/Maxlego08/zEssentials/issues/33) 
+- Upgrade to [Sarah](https://github.com/Maxlego08/Sarah/) version 1.10
+- Fixed issue where savings by default could not be removed [#34](https://github.com/Maxlego08/zEssentials/issues/34). You must update your configuration like this
+Before
+````yaml
+economies:
+  money:
+    display-name: Money
+````
+After
+````yaml
+economies:
+  - name: money
+    display-name: Money
+````
+- Fixed permissions for night vision and vault commands
+- Added command ``/sudo`` [#36](https://github.com/Maxlego08/zEssentials/issues/36)
+
 # 1.0.0.3
 
 - Change ``AsyncPlayerPreLoginEvent`` to ``PlayerLoginEvent``